Update the Grobid parser to operate as a blob parser, move logic from grobid loader and remove
