+++ docs/extras/ecosystem/integrations/grobid.mdx
@@ -20,12 +20,25 @@ get_ipython().system_raw('nohup ./gradlew run > grobid.log 2>&1 &')
 
 You can now use the GrobidParser to produce documents
 ```python
-from langchain.documentloaders import GrobidLoader
+from langchain.document_loaders.parsers import GrobidParser
+from langchain.document_loaders.generic import GenericLoader
 
 #Produce chunks from article paragraphs
-chunks=GrobidLoader(file_path,segment_sentences=False)
+loader = GenericLoader.from_filesystem(
+    "/Users/31treehaus/Desktop/Papers/",
+    glob="*",
+    suffixes=[".pdf"],
+    parser= GrobidParser(segment_sentences=False)
+)
+docs = loader.load()
 
 #Produce chunks from article sentences
-chunks=GrobidLoader(file_path,segment_sentences=True)
+loader = GenericLoader.from_filesystem(
+    "/Users/31treehaus/Desktop/Papers/",
+    glob="*",
+    suffixes=[".pdf"],
+    parser= GrobidParser(segment_sentences=True)
+)
+docs = loader.load()
 ```
 Chunk metadata will include bboxes although these are a bit funky to parse, see https://grobid.readthedocs.io/en/latest/Coordinates-in-PDF/
