+++ langchain/document_loaders/parsers/__init__.py
@@ -1,4 +1,5 @@
 from langchain.document_loaders.parsers.audio import OpenAIWhisperParser
+from langchain.document_loaders.parsers.grobid import GrobidParser
 from langchain.document_loaders.parsers.html import BS4HTMLParser
 from langchain.document_loaders.parsers.language import LanguageParser
 from langchain.document_loaders.parsers.pdf import (
@@ -11,6 +12,7 @@ from langchain.document_loaders.parsers.pdf import (
 
 __all__ = [
     "BS4HTMLParser",
+    "GrobidParser",
     "LanguageParser",
     "OpenAIWhisperParser",
     "PDFMinerParser",
