+++ tests/unit_tests/document_loaders/parsers/test_public_api.py
@@ -5,6 +5,7 @@ def test_parsers_public_api_correct() -> None:
     """Test public API of parsers for breaking changes."""
     assert set(__all__) == {
         "BS4HTMLParser",
+        "GrobidParser",
         "LanguageParser",
         "OpenAIWhisperParser",
         "PyPDFParser",
